Provide v1 seedfile when v2 size matches
