+++ internal/pkg/agent/vault/seed_fips_test.go
@@ -12,6 +12,8 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/google/go-cmp/cmp"
+
 	"github.com/elastic/elastic-agent/internal/pkg/agent/vault/aesgcm"
 )
 
@@ -35,5 +37,41 @@ func TestGetSeedFailsV1File(t *testing.T) {
 	_, _, err = getSeed(dir)
 	if !errors.Is(err, os.ErrNotExist) {
 		t.Fatalf("FIPS mode must not read v1 seeds. expected error %v to be os.ErrNotExist", err)
+	}
+}
+
+func TestCreateSeedIfNotExists(t *testing.T) {
+	dir := t.TempDir()
+
+	fp := filepath.Join(dir, seedFile)
+	fpV2 := filepath.Join(dir, seedFileV2)
+
+	if _, err := os.Stat(fp); !errors.Is(err, os.ErrNotExist) {
+		t.Fatal(err)
+	}
+	if _, err := os.Stat(fpV2); !errors.Is(err, os.ErrNotExist) {
+		t.Fatal(err)
+	}
+
+	b, saltSize, err := createSeedIfNotExists(dir)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// V2 file should exist
+	if _, err := os.Stat(fpV2); err != nil {
+		t.Fatal(err)
+	}
+	// V1 file should not exist
+	if _, err := os.Stat(fp); !errors.Is(err, os.ErrNotExist) {
+		t.Fatal(err)
+	}
+
+	diff := cmp.Diff(int(aesgcm.AES256), len(b))
+	if diff != "" {
+		t.Error(diff)
+	}
+	if saltSize != defaultSaltSize {
+		t.Errorf("expected salt size: %d got: %d", defaultSaltSize, saltSize)
 	}
 }
