+++ internal/pkg/agent/vault/seed_nofips_test.go
@@ -45,3 +45,39 @@ func TestGetSeedReturnsV1File(t *testing.T) {
 		t.Errorf("expected salt size: %d got: %d", saltSizeV1, saltSize)
 	}
 }
+
+func TestCreateSeedIfNotExists(t *testing.T) {
+	dir := t.TempDir()
+
+	fp := filepath.Join(dir, seedFile)
+	fpV2 := filepath.Join(dir, seedFileV2)
+
+	if _, err := os.Stat(fp); !errors.Is(err, os.ErrNotExist) {
+		t.Fatal(err)
+	}
+	if _, err := os.Stat(fpV2); !errors.Is(err, os.ErrNotExist) {
+		t.Fatal(err)
+	}
+
+	b, saltSize, err := createSeedIfNotExists(dir)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// V2 file should exist
+	if _, err := os.Stat(fpV2); err != nil {
+		t.Fatal(err)
+	}
+	// V1 file should exist
+	if _, err := os.Stat(fp); err != nil {
+		t.Fatal(err)
+	}
+
+	diff := cmp.Diff(int(aesgcm.AES256), len(b))
+	if diff != "" {
+		t.Error(diff)
+	}
+	if saltSize != defaultSaltSize {
+		t.Errorf("expected salt size: %d got: %d", defaultSaltSize, saltSize)
+	}
+}
