+++ internal/pkg/agent/vault/seed_test.go
@@ -55,10 +55,7 @@ func TestGetSeed(t *testing.T) {
 	if _, err := os.Stat(fpV2); err != nil {
 		t.Fatal(err)
 	}
-	// V1 file should not exist
-	if _, err := os.Stat(fp); !errors.Is(err, os.ErrNotExist) {
-		t.Fatal(err)
-	}
+	// V1 file is checked as a part of TestCreateSeedIfNotExists
 
 	diff := cmp.Diff(int(aesgcm.AES256), len(b))
 	if diff != "" {
@@ -80,42 +77,6 @@ func TestGetSeed(t *testing.T) {
 	}
 	if gotSaltSize != saltSize {
 		t.Errorf("got salt size %d does not match written salt size: %d", gotSaltSize, saltSize)
-	}
-}
-
-func TestCreateSeedIfNotExists(t *testing.T) {
-	dir := t.TempDir()
-
-	fp := filepath.Join(dir, seedFile)
-	fpV2 := filepath.Join(dir, seedFileV2)
-
-	if _, err := os.Stat(fp); !errors.Is(err, os.ErrNotExist) {
-		t.Fatal(err)
-	}
-	if _, err := os.Stat(fpV2); !errors.Is(err, os.ErrNotExist) {
-		t.Fatal(err)
-	}
-
-	b, saltSize, err := createSeedIfNotExists(dir)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// V2 file should exist
-	if _, err := os.Stat(fpV2); err != nil {
-		t.Fatal(err)
-	}
-	// V1 file should not exist
-	if _, err := os.Stat(fp); !errors.Is(err, os.ErrNotExist) {
-		t.Fatal(err)
-	}
-
-	diff := cmp.Diff(int(aesgcm.AES256), len(b))
-	if diff != "" {
-		t.Error(diff)
-	}
-	if saltSize != defaultSaltSize {
-		t.Errorf("expected salt size: %d got: %d", defaultSaltSize, saltSize)
 	}
 }
